Update 2
Added onmouseover and onmouseout functionality, hiding the results when the mouse is not hovering over the answers. Also implemented clicking and locking in the answers, and created the threeofkind, fourofkind and yahtzee functions. Still need to fixc calculations, fix the new game function and finally complete the total sum calculation.